# example-node/api/swagger/swagger.yaml
...
swagger: "2.0"
...
host: localhost:3000
...
info:
  title: DMC example for Node.js
  description: Design-based Management Console.
  version: 0.0.1

#################################################
# schemes
# - APIを提供する通信プロトコルのリスト
#################################################
schemes:
- https

#################################################
# definitions
# - リクエストやレスポンスのレイアウトの定義
# - paths内から $ref で参照できる
#################################################
definitions:
...
  error:
    description: Error response media type
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1502172152
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1502172152
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    type: object
...
paths:
...
  /user/download/csv:
    get:
      x-swagger-router-controller: user
      description: ユーザ一覧をcsvファイルでダウンロード
      operationId: user#download
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      security:
      - jwt:
        - api:access
      summary: ユーザ一覧ダウンロード
      tags:
      - user
...
securityDefinitions:
  jwt:
    type: oauth2
    authorizationUrl: /signin
    flow: implicit
    scopes:
      api:access: API access

#################################################
# produces
# - APIが提供するMIMEタイプの一覧
#################################################
produces:
  - application/json
  - text/plain

#################################################
# consumes
# - APIが受け付けるMIMEタイプの一覧
#################################################
consumes:
  - application/json
  - text/plain